+++ contracts/TripleSlopeRateModel.sol
@@ -4,13 +4,20 @@ import "./InterestRateModel.sol";
 import "./SafeMath.sol";
 
 /**
-  * @title CREAM's TripleSlopeRateModel Contract
-  * @author C.R.E.A.M. Finance
-  */
+ * @title CREAM's TripleSlopeRateModel Contract
+ * @author C.R.E.A.M. Finance
+ */
 contract TripleSlopeRateModel is InterestRateModel {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
-    event NewInterestParams(uint baseRatePerBlock, uint multiplierPerBlock, uint jumpMultiplierPerBlock, uint kink1, uint kink2, uint roof);
+    event NewInterestParams(
+        uint256 baseRatePerBlock,
+        uint256 multiplierPerBlock,
+        uint256 jumpMultiplierPerBlock,
+        uint256 kink1,
+        uint256 kink2,
+        uint256 roof
+    );
 
     /**
      * @notice The address of the owner, i.e. the Timelock contract, which can update parameters directly
@@ -20,42 +27,42 @@ contract TripleSlopeRateModel is InterestRateModel {
     /**
      * @notice The approximate number of blocks per year that is assumed by the interest rate model
      */
-    uint public constant blocksPerYear = 2102400;
+    uint256 public constant blocksPerYear = 2102400;
 
     /**
      * @notice The minimum roof value used for calculating borrow rate.
      */
-    uint internal constant minRoofValue = 1e18;
+    uint256 internal constant minRoofValue = 1e18;
 
     /**
      * @notice The multiplier of utilization rate that gives the slope of the interest rate
      */
-    uint public multiplierPerBlock;
+    uint256 public multiplierPerBlock;
 
     /**
      * @notice The base interest rate which is the y-intercept when utilization rate is 0
      */
-    uint public baseRatePerBlock;
+    uint256 public baseRatePerBlock;
 
     /**
      * @notice The multiplierPerBlock after hitting a specified utilization point
      */
-    uint public jumpMultiplierPerBlock;
+    uint256 public jumpMultiplierPerBlock;
 
     /**
      * @notice The utilization point at which the interest rate is fixed
      */
-    uint public kink1;
+    uint256 public kink1;
 
     /**
      * @notice The utilization point at which the jump multiplier is applied
      */
-    uint public kink2;
+    uint256 public kink2;
 
     /**
      * @notice The utilization point at which the rate is fixed
      */
-    uint public roof;
+    uint256 public roof;
 
     /**
      * @notice Construct an interest rate model
@@ -67,10 +74,18 @@ contract TripleSlopeRateModel is InterestRateModel {
      * @param roof_ The utilization point at which the borrow rate is fixed
      * @param owner_ The address of the owner, i.e. the Timelock contract (which has the ability to update parameters directly)
      */
-    constructor(uint baseRatePerYear, uint multiplierPerYear, uint jumpMultiplierPerYear, uint kink1_, uint kink2_, uint roof_, address owner_) public {
+    constructor(
+        uint256 baseRatePerYear,
+        uint256 multiplierPerYear,
+        uint256 jumpMultiplierPerYear,
+        uint256 kink1_,
+        uint256 kink2_,
+        uint256 roof_,
+        address owner_
+    ) public {
         owner = owner_;
 
-        updateTripleRateModelInternal(baseRatePerYear,  multiplierPerYear, jumpMultiplierPerYear, kink1_, kink2_, roof_);
+        updateTripleRateModelInternal(baseRatePerYear, multiplierPerYear, jumpMultiplierPerYear, kink1_, kink2_, roof_);
     }
 
     /**
@@ -82,7 +97,14 @@ contract TripleSlopeRateModel is InterestRateModel {
      * @param kink2_ The utilization point at which the jump multiplier is applied
      * @param roof_ The utilization point at which the borrow rate is fixed
      */
-    function updateTripleRateModel(uint baseRatePerYear, uint multiplierPerYear, uint jumpMultiplierPerYear, uint kink1_, uint kink2_, uint roof_) external {
+    function updateTripleRateModel(
+        uint256 baseRatePerYear,
+        uint256 multiplierPerYear,
+        uint256 jumpMultiplierPerYear,
+        uint256 kink1_,
+        uint256 kink2_,
+        uint256 roof_
+    ) external {
         require(msg.sender == owner, "only the owner may call this function.");
 
         updateTripleRateModelInternal(baseRatePerYear, multiplierPerYear, jumpMultiplierPerYear, kink1_, kink2_, roof_);
@@ -95,13 +117,17 @@ contract TripleSlopeRateModel is InterestRateModel {
      * @param reserves The amount of reserves in the market (currently unused)
      * @return The utilization rate as a mantissa between [0, 1e18]
      */
-    function utilizationRate(uint cash, uint borrows, uint reserves) public view returns (uint) {
+    function utilizationRate(
+        uint256 cash,
+        uint256 borrows,
+        uint256 reserves
+    ) public view returns (uint256) {
         // Utilization rate is 0 when there are no borrows
         if (borrows == 0) {
             return 0;
         }
 
-        uint util = borrows.mul(1e18).div(cash.add(borrows).sub(reserves));
+        uint256 util = borrows.mul(1e18).div(cash.add(borrows).sub(reserves));
         // If the utilization is above the roof, cap it.
         if (util > roof) {
             util = roof;
@@ -116,16 +142,20 @@ contract TripleSlopeRateModel is InterestRateModel {
      * @param reserves The amount of reserves in the market
      * @return The borrow rate percentage per block as a mantissa (scaled by 1e18)
      */
-    function getBorrowRate(uint cash, uint borrows, uint reserves) public view returns (uint) {
-        uint util = utilizationRate(cash, borrows, reserves);
+    function getBorrowRate(
+        uint256 cash,
+        uint256 borrows,
+        uint256 reserves
+    ) public view returns (uint256) {
+        uint256 util = utilizationRate(cash, borrows, reserves);
 
         if (util <= kink1) {
             return util.mul(multiplierPerBlock).div(1e18).add(baseRatePerBlock);
         } else if (util <= kink2) {
             return kink1.mul(multiplierPerBlock).div(1e18).add(baseRatePerBlock);
         } else {
-            uint normalRate = kink1.mul(multiplierPerBlock).div(1e18).add(baseRatePerBlock);
-            uint excessUtil = util.sub(kink2);
+            uint256 normalRate = kink1.mul(multiplierPerBlock).div(1e18).add(baseRatePerBlock);
+            uint256 excessUtil = util.sub(kink2);
             return excessUtil.mul(jumpMultiplierPerBlock).div(1e18).add(normalRate);
         }
     }
@@ -138,10 +168,15 @@ contract TripleSlopeRateModel is InterestRateModel {
      * @param reserveFactorMantissa The current reserve factor for the market
      * @return The supply rate percentage per block as a mantissa (scaled by 1e18)
      */
-    function getSupplyRate(uint cash, uint borrows, uint reserves, uint reserveFactorMantissa) public view returns (uint) {
-        uint oneMinusReserveFactor = uint(1e18).sub(reserveFactorMantissa);
-        uint borrowRate = getBorrowRate(cash, borrows, reserves);
-        uint rateToPool = borrowRate.mul(oneMinusReserveFactor).div(1e18);
+    function getSupplyRate(
+        uint256 cash,
+        uint256 borrows,
+        uint256 reserves,
+        uint256 reserveFactorMantissa
+    ) public view returns (uint256) {
+        uint256 oneMinusReserveFactor = uint256(1e18).sub(reserveFactorMantissa);
+        uint256 borrowRate = getBorrowRate(cash, borrows, reserves);
+        uint256 rateToPool = borrowRate.mul(oneMinusReserveFactor).div(1e18);
         return utilizationRate(cash, borrows, reserves).mul(rateToPool).div(1e18);
     }
 
@@ -154,7 +189,14 @@ contract TripleSlopeRateModel is InterestRateModel {
      * @param kink2_ The utilization point at which the jump multiplier is applied
      * @param roof_ The utilization point at which the borrow rate is fixed
      */
-    function updateTripleRateModelInternal(uint baseRatePerYear, uint multiplierPerYear, uint jumpMultiplierPerYear, uint kink1_, uint kink2_, uint roof_) internal {
+    function updateTripleRateModelInternal(
+        uint256 baseRatePerYear,
+        uint256 multiplierPerYear,
+        uint256 jumpMultiplierPerYear,
+        uint256 kink1_,
+        uint256 kink2_,
+        uint256 roof_
+    ) internal {
         require(kink1_ <= kink2_, "kink1 must less than or equal to kink2");
         require(roof_ >= minRoofValue, "invalid roof value");
