+++ spec/certora/contracts/CDaiDelegateCertora.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.5.16;
 import "../../../contracts/CDaiDelegate.sol";
 
 contract CDaiDelegateCertora is CDaiDelegate {
-    function getCashOf(address account) public view returns (uint) {
+    function getCashOf(address account) public view returns (uint256) {
         return EIP20Interface(underlying).balanceOf(account);
     }
 }
